add tests for metadata table
