+++ pkg/osquery/tables/gsettings/gsettings.go
@@ -4,7 +4,6 @@ import (
 	"bufio"
 	"bytes"
 	"context"
-	"fmt"
 	"io/ioutil"
 	"os"
 	"os/exec"
@@ -15,24 +14,19 @@ import (
 	"github.com/go-kit/kit/log/level"
 	"github.com/kolide/launcher/pkg/dataflatten"
 	"github.com/kolide/launcher/pkg/osquery/tables/dataflattentable"
-	"github.com/kolide/launcher/pkg/osquery/tables/tablehelpers"
 	osquery "github.com/kolide/osquery-go"
 	"github.com/kolide/osquery-go/plugin/table"
 	"github.com/pkg/errors"
 )
 
 const gsettingsPath = "/usr/bin/gsettings"
+
 const allowedCharacters = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789_-."
 
 type GsettingsValues struct {
 	client   *osquery.ExtensionManagerClient
 	logger   log.Logger
 	getBytes func(ctx context.Context, buf *bytes.Buffer) error
-}
-
-type GsettingsMetadata struct {
-	client *osquery.ExtensionManagerClient
-	logger log.Logger
 }
 
 // Settings returns a table plugin for querying setting values from the
@@ -134,273 +128,6 @@ func (t *GsettingsValues) flatten(buffer *bytes.Buffer) ([]dataflatten.Row, erro
 		rows = append(rows, row)
 	}
 	return rows, nil
-}
-
-// Metadata returns a table plugin for querying metadata about specific keys in
-// specific schemas
-func Metadata(client *osquery.ExtensionManagerClient, logger log.Logger) *table.Plugin {
-
-	columns := []table.ColumnDefinition{
-		// TODO: maybe need to add 'path' for relocatable schemas..
-		table.TextColumn("schema"),
-		table.TextColumn("key"),
-		table.TextColumn("description"),
-		table.TextColumn("type"),
-	}
-
-	t := &GsettingsMetadata{
-		client: client,
-		logger: logger,
-	}
-
-	return table.NewPlugin("kolide_gsettings_metadata", columns, t.generate)
-}
-
-func (t *GsettingsMetadata) generate(ctx context.Context, queryContext table.QueryContext) ([]map[string]string, error) {
-
-	var results []map[string]string
-
-	for _, schema := range tablehelpers.GetConstraints(queryContext, "schema", tablehelpers.WithAllowedCharacters(allowedCharacters)) {
-		descriptions, err := t.gsettingsDescribeForSchema(ctx, schema, "")
-		if err != nil {
-			level.Info(t.logger).Log(
-				"msg", "error describing keys for schema",
-				"schema", schema,
-				"err", err,
-			)
-			continue
-		}
-		for _, d := range descriptions {
-			row := map[string]string{
-				"description": d.Description,
-				"type":        d.Type,
-				"schema":      schema,
-				"key":         d.Key,
-			}
-			results = append(results, row)
-		}
-
-	}
-
-	return results, nil
-}
-
-type datatype struct {
-	raw string
-}
-
-type keyDescription struct {
-	Description string
-	Type        string
-	Key         string
-}
-
-func (t *GsettingsMetadata) gsettingsDescribeForSchema(ctx context.Context, schema, key string) ([]keyDescription, error) {
-	var descriptions []keyDescription
-	var keys []string
-	var err error
-
-	if key != "" {
-		keys = append(keys, key)
-	} else {
-		keys, err = t.listKeys(ctx, schema, t.logger)
-		if err != nil {
-			return descriptions, errors.Wrap(err, "fetching keys to describe")
-		}
-	}
-
-	for _, k := range keys {
-		desc, err := t.describeKey(ctx, schema, k)
-		if err != nil {
-			level.Info(t.logger).Log(
-				"msg", "error describing key",
-				"key", key,
-				"schema", schema,
-				"err", err,
-			)
-			continue
-		}
-		descriptions = append(descriptions, desc)
-	}
-
-	return descriptions, nil
-}
-
-func (t *GsettingsMetadata) listKeys(ctx context.Context, schema string, l log.Logger) ([]string, error) {
-	ctx, cancel := context.WithTimeout(ctx, 3*time.Second)
-	defer cancel()
-	var keys []string
-
-	cmd := exec.CommandContext(ctx, "/usr/bin/gsettings", "list-keys", schema)
-	dir, err := ioutil.TempDir("", "osq-gsettings-list-keys")
-	if err != nil {
-		return keys, errors.Wrap(err, "mktemp")
-	}
-	defer os.RemoveAll(dir)
-
-	if err := os.Chmod(dir, 0755); err != nil {
-		return keys, errors.Wrap(err, "chmod")
-	}
-	cmd.Dir = dir
-	stdout, stderr := new(bytes.Buffer), new(bytes.Buffer)
-	cmd.Stdout, cmd.Stderr = stdout, stderr
-
-	if err := cmd.Run(); err != nil {
-		level.Info(t.logger).Log(
-			"msg", "Error running gsettings list-keys",
-			"schema", schema,
-			"stderr", strings.TrimSpace(stderr.String()),
-			"stdout", strings.TrimSpace(stdout.String()),
-			"err", err,
-		)
-		return keys, errors.Wrap(err, "running gsettings list-keys")
-	}
-
-	scanner := bufio.NewScanner(stdout)
-
-	for scanner.Scan() {
-		line := scanner.Text()
-		if line == "" {
-			continue
-		}
-
-		keys = append(keys, line)
-	}
-
-	if err := scanner.Err(); err != nil {
-		level.Info(t.logger).Log("msg", "scanner error", "err", err)
-	}
-
-	return keys, nil
-}
-
-func (t *GsettingsMetadata) describeKey(ctx context.Context, schema, key string) (keyDescription, error) {
-	ctx, cancel := context.WithTimeout(ctx, 3*time.Second)
-	defer cancel()
-	desc := keyDescription{Key: key}
-	cmd := exec.CommandContext(ctx, "/usr/bin/gsettings", "describe", schema, key)
-	dir, err := ioutil.TempDir("", "osq-gsettings-desc")
-	if err != nil {
-		return keyDescription{}, errors.Wrap(err, "mktemp")
-	}
-	defer os.RemoveAll(dir)
-
-	if err := os.Chmod(dir, 0755); err != nil {
-		return keyDescription{}, errors.Wrap(err, "chmod")
-	}
-	cmd.Dir = dir
-	stdout, stderr := new(bytes.Buffer), new(bytes.Buffer)
-	cmd.Stdout, cmd.Stderr = stdout, stderr
-
-	if err := cmd.Run(); err != nil {
-		level.Debug(t.logger).Log(
-			"msg", "Error running gsettings describe",
-			"key", key,
-			"schema", schema,
-			"stderr", strings.TrimSpace(stderr.String()),
-			"stdout", strings.TrimSpace(stdout.String()),
-			"err", err,
-		)
-		return keyDescription{}, errors.Wrap(err, "running gsettings describe")
-	}
-
-	desc.Description = strings.TrimSpace(stdout.String())
-	datatype, err := t.getType(ctx, key, schema)
-	if err != nil {
-		return desc, errors.Wrap(err, "discerning key's type")
-	}
-	desc.Type = datatype
-
-	return desc, nil
-}
-
-func (t *GsettingsMetadata) getType(ctx context.Context, key, schema string) (string, error) {
-	cmd := exec.CommandContext(ctx, "/usr/bin/gsettings", "range", schema, key)
-	dir, err := ioutil.TempDir("", "osq-gsettings-range")
-	if err != nil {
-		return "", errors.Wrap(err, "mktemp")
-	}
-	defer os.RemoveAll(dir)
-
-	if err := os.Chmod(dir, 0755); err != nil {
-		return "", errors.Wrap(err, "chmod")
-	}
-	cmd.Dir = dir
-	stdout, stderr := new(bytes.Buffer), new(bytes.Buffer)
-	cmd.Stdout, cmd.Stderr = stdout, stderr
-
-	if err := cmd.Run(); err != nil {
-		level.Debug(t.logger).Log(
-			"msg", "Error running gsettings range",
-			"key", key,
-			"schema", schema,
-			"stderr", strings.TrimSpace(stderr.String()),
-			"stdout", strings.TrimSpace(stdout.String()),
-			"err", err,
-		)
-		return "", errors.Wrap(err, "running gsettings range")
-	}
-
-	result := strings.ReplaceAll(strings.TrimSpace(stdout.String()), "\n", " ")
-
-	// enum types need special formatting to distinguish the type (enum) from
-	// the possible values
-	if strings.HasPrefix(result, "enum") {
-		s := strings.TrimPrefix(result, "enum")
-		vals := strings.Split(s, " ")
-		return fmt.Sprintf("enum: [ %s ]", strings.Join(vals, ",")), nil
-	}
-
-	// 'range' datatypes also need special handling
-	if strings.HasPrefix(result, "range ") {
-		s := strings.TrimPrefix(result, "range ")
-		parts := strings.Split(s, " ")
-
-		typ := convertType(parts[0])
-		var scope string
-		if len(parts) > 2 {
-			scope = fmt.Sprintf(" (%v to %v)", parts[1], parts[2])
-		}
-
-		return fmt.Sprintf("%s%s", typ, scope), nil
-	}
-
-	return convertType(result), nil
-}
-
-// this feels.. not idiomatic
-var gvariantMapping = map[string]string{
-	"b": "bool",
-	"n": "int16",
-	"q": "uint16",
-	"i": "int32",
-	"u": "uint32",
-	"x": "int64",
-	"t": "uint64",
-	"d": "double",
-	"s": "string",
-	"a": "array",
-}
-
-// convertType returns a string describing the GVariantType corresponding to the
-// GVariant-formatted type string. see  https://developer.gnome.org/glib/unstable/glib-GVariantType.html
-// for documentation. Note that not all types listed in the documentation above
-// are supported, for example:
-//  - tuples (e.g. tuple of 2 strings `(ss)`)
-//  - nested types (e.g.// array of tuples: `a(ss)`)
-// and other complex types are not supported.
-func convertType(typ string) string {
-	typ = strings.Replace(typ, "type ", "", 1) // remove any leading 'type ', eg 'type b'
-	var prefix string
-	if strings.HasPrefix(typ, "a") {
-		typ = typ[1:]
-		prefix = "array of "
-	}
-	primitive_typ, ok := gvariantMapping[typ]
-	if !ok {
-		return "other"
-	}
-	return fmt.Sprintf("%s%s", prefix, primitive_typ)
 }
 
 func (t *GsettingsValues) parse(input *bytes.Buffer) []map[string]string {
